Moved model loading function out of CameraView to its own class so Marvin doesn't kill me.
